+++ circuitbreaker_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/cenkalti/backoff"
+	"github.com/cenk/backoff"
 	"github.com/facebookgo/clock"
 )
 
@@ -97,6 +97,44 @@ func TestBreakerEvents(t *testing.T) {
 	cb.Fail()
 	if e := <-events; e != BreakerFail {
 		t.Fatalf("expected to receive a fail event, got %d", e)
+	}
+}
+
+func TestAddRemoveListener(t *testing.T) {
+	c := clock.NewMock()
+	cb := NewBreaker()
+	cb.Clock = c
+	events := make(chan ListenerEvent, 100)
+	cb.AddListener(events)
+
+	cb.Trip()
+	if e := <-events; e.Event != BreakerTripped {
+		t.Fatalf("expected to receive a trip event, got %v", e)
+	}
+
+	c.Add(cb.nextBackOff + 1)
+	cb.Ready()
+	if e := <-events; e.Event != BreakerReady {
+		t.Fatalf("expected to receive a breaker ready event, got %v", e)
+	}
+
+	cb.Reset()
+	if e := <-events; e.Event != BreakerReset {
+		t.Fatalf("expected to receive a reset event, got %v", e)
+	}
+
+	cb.Fail()
+	if e := <-events; e.Event != BreakerFail {
+		t.Fatalf("expected to receive a fail event, got %v", e)
+	}
+
+	cb.RemoveListener(events)
+	cb.Reset()
+	select {
+	case e := <-events:
+		t.Fatalf("after removing listener, should not receive reset event; got %v", e)
+	default:
+		// Expected.
 	}
 }
 
@@ -248,21 +286,37 @@ func TestThresholdBreakerResets(t *testing.T) {
 }
 
 func TestTimeoutBreaker(t *testing.T) {
+	wait := make(chan struct{})
+
 	c := clock.NewMock()
 	called := int32(0)
+
 	circuit := func() error {
+		wait <- struct{}{}
 		atomic.AddInt32(&called, 1)
-		c.Add(time.Millisecond)
+		<-wait
 		return nil
 	}
 
 	cb := NewThresholdBreaker(1)
 	cb.Clock = c
-	err := cb.Call(circuit, time.Millisecond)
+
+	errc := make(chan error)
+	go func() { errc <- cb.Call(circuit, time.Millisecond) }()
+
+	<-wait
+	c.Add(time.Millisecond * 3)
+	wait <- struct{}{}
+
+	err := <-errc
 	if err == nil {
 		t.Fatal("expected timeout breaker to return an error")
 	}
-	cb.Call(circuit, time.Millisecond)
+
+	go cb.Call(circuit, time.Millisecond)
+	<-wait
+	c.Add(time.Millisecond * 3)
+	wait <- struct{}{}
 
 	if !cb.Tripped() {
 		t.Fatal("expected timeout breaker to be open")