+++ window.go
@@ -4,6 +4,8 @@ import (
 	"container/ring"
 	"sync"
 	"time"
+
+	"github.com/facebookgo/clock"
 )
 
 var (
@@ -45,6 +47,7 @@ type window struct {
 	bucketTime time.Duration
 	bucketLock sync.RWMutex
 	lastAccess time.Time
+	clock      clock.Clock
 }
 
 // newWindow creates a new window. windowTime is the time covering the entire
@@ -58,8 +61,15 @@ func newWindow(windowTime time.Duration, windowBuckets int) *window {
 		buckets = buckets.Next()
 	}
 
+	clock := clock.New()
+
 	bucketTime := time.Duration(windowTime.Nanoseconds() / int64(windowBuckets))
-	return &window{buckets: buckets, bucketTime: bucketTime, lastAccess: time.Now()}
+	return &window{
+		buckets:    buckets,
+		bucketTime: bucketTime,
+		clock:      clock,
+		lastAccess: clock.Now(),
+	}
 }
 
 // Fail records a failure in the current bucket.
@@ -143,12 +153,22 @@ func (w *window) Reset() {
 func (w *window) getLatestBucket() *bucket {
 	var b *bucket
 	b = w.buckets.Value.(*bucket)
+	elapsed := w.clock.Now().Sub(w.lastAccess)
 
-	if time.Since(w.lastAccess) > w.bucketTime {
-		w.buckets = w.buckets.Next()
-		b = w.buckets.Value.(*bucket)
-		b.Reset()
-		w.lastAccess = time.Now()
+	if elapsed > w.bucketTime {
+		// Reset the buckets between now and number of buckets ago. If
+		// that is more that the existing buckets, reset all.
+		for i := 0; i < w.buckets.Len(); i++ {
+			w.buckets = w.buckets.Next()
+			b = w.buckets.Value.(*bucket)
+			b.Reset()
+			elapsed = time.Duration(int64(elapsed) - int64(w.bucketTime))
+			if elapsed < w.bucketTime {
+				// Done resetting buckets.
+				break
+			}
+		}
+		w.lastAccess = w.clock.Now()
 	}
 	return b
 }